+++ apps/api/src/routes/index.ts
@@ -1,23 +1,34 @@
 import { Router } from "express";
 import authenticate from "../middleware/jwt-auth-middleware";
-import CompanyRoutes from "./CompanyRoutes";
+import CompanyController from "../controllers/CompanyController";
+import ProxyController from "../controllers/ProxyController";
+import AuthController from "../controllers/AuthController";
 
 const router = Router();
 
+// Auth token
+router.post("/im/auth/token", AuthController.token);
+
 // Signup & Login
-router.post("/companies/signup", CompanyRoutes.signup);
-router.post("/companies/login", CompanyRoutes.login);
+router.post("/directory/companies/signup", CompanyController.signup);
+router.post("/directory/companies/login", CompanyController.login);
 
 // Password reset
-router.post("/companies/forgot-password", CompanyRoutes.forgotPassword);
-router.post("/companies/reset-password", CompanyRoutes.resetPassword);
-router.get("/companies/verify-reset-token", CompanyRoutes.verifyResetToken);
+router.post("/directory/companies/forgot-password", CompanyController.forgotPassword);
+router.post("/directory/companies/reset-password", CompanyController.resetPassword);
+router.get("/directory/companies/verify-reset-token", CompanyController.verifyResetToken);
 
-// Private routes
-router.get("/companies/my-profile", authenticate, CompanyRoutes.myProfile);
-router.get("/companies/profile", authenticate, CompanyRoutes.getCompany);
-router.get("/companies/search", authenticate, CompanyRoutes.searchCompanies);
-router.post("/companies/create-connection-request", authenticate, CompanyRoutes.createConnectionRequest);
-router.post("/companies/connection-request-action", authenticate, CompanyRoutes.connectionRequestAction);
+// Company profile and search
+router.get("/directory/companies/my-profile", authenticate, CompanyController.myProfile);
+router.get("/directory/companies/profile", authenticate, CompanyController.getCompany);
+router.get("/directory/companies/search", authenticate, CompanyController.searchCompanies);
+router.post("/directory/companies/create-connection-request", authenticate, CompanyController.createConnectionRequest);
+router.post("/directory/companies/connection-request-action", authenticate, CompanyController.connectionRequestAction);
+
+// Proxy routes to Conformance service
+router.post("/proxy/test", authenticate, ProxyController.runTestCases);
+router.get("/proxy/test-results", authenticate, ProxyController.getTestResults);
+router.get("/proxy/test-runs", authenticate, ProxyController.getOrSearchTestRuns);
+
 
 export default router;
